Add spring and factorialize
