+++ src/main/java/org/eolang/benchmark/Main.java
@@ -23,14 +23,35 @@
  */
 package org.eolang.benchmark;
 
-public class Main {
-    public static void main(String... args) {
-        long total = Long.parseLong(args[0]);
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * Main class.
+ * We expect that the usage of Factorial objects in this class will be replaced.
+ * @since 0.2
+ */
+@RestController
+public final class Main {
+
+    @GetMapping("/factorial")
+    public String factorial(
+        @RequestParam(value = "rounds", defaultValue = "1") String prounds,
+        @RequestParam(value = "value", defaultValue = "10") String pvalue
+    ) {
+        long rounds = Long.parseLong(prounds);
+        int value = Integer.parseInt(pvalue);
         long sum = 0L;
         long start = System.currentTimeMillis();
-        for (long i = 0; i < total; ++i) {
-            sum += new A(42).get();
+        for (long i = 0; i < rounds; ++i) {
+            sum += new Factorial(value).get();
         }
-        System.out.printf("sum=%d time=%d\n", sum, System.currentTimeMillis() - start);
+        final String response = String.format(
+            "sum=%d time=%d\n", sum, System.currentTimeMillis() - start
+        );
+        System.out.println(response);
+        return response;
     }
+
 }
